kpi-dashboard/backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
pandas
openpyxl
flask-cors
scikit-learn
numpy
chromadb
sentence-transformers
anthropic
faiss-cpu
openai
python-dotenv
qdrant-client 